requirements.txt
amqp==2.5.2
arrow==0.15.4
astroid==2.3.3
aws-requests-auth==0.4.2
bcrypt==3.1.7
blinker==1.4
boto3==1.11.9
botocore==1.14.9
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coloredlogs==10.0
deepmerge==0.1.0
docutils==0.15.2
ftfy==5.6
humanfriendly==6.1
idna==2.8
importlib-metadata==1.4.0
isort==4.3.21
jmespath==0.9.4
kombu==4.6.7
lazy-object-proxy==1.4.3
mailchecker==3.3.4
mccabe==0.6.1
mongoengine==0.19.1
phonenumbers==8.11.3
pycparser==2.19
pydash==4.7.6
pylint==2.4.4
pymongo==3.9.0
python-benedict==0.16.0
python-dateutil==2.8.1
python-dotenv==0.10.5
python-slugify==4.0.0
PyYAML==5.3
redis==3.3.11
requests==2.22.0
ruamel.yaml==0.16.7
ruamel.yaml.clib==0.2.0
s3transfer==0.3.2
six==1.13.0
text-unidecode==1.3
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
wrapt==1.11.2
xmltodict==0.12.0
zipp==2.1.0
../campus-shared-models
../campus-libs
